Add extend extension attributes to <store_url>/V1/orders request
