+++ etc/webapi_rest/di.xml
@@ -17,4 +17,9 @@
                 type="Extend\Warranty\Plugin\Api\OrderItemRepositoryInterfacePlugin" sortOrder="10" />
     </type>
 
+    <type name="Magento\Sales\Api\OrderRepositoryInterface">
+        <plugin name="extend_warranty_plugin_api_order_repository_interface_plugin"
+                type="Extend\Warranty\Plugin\Api\OrderRepositoryInterfacePlugin" sortOrder="10" />
+    </type>
+
 </config>
